Propagate parameter events unidirectionally through the native parameter store
This essentially implements the flux pattern where the sliders dispatch
events that the AudioProcessorValueTreeState picks up. In response,
parameter listeners fire and dispatch the new values back to the script
environment, where the react components learn and update their internal
state

This includes as well pushing string values into the JavaScript side so
that the labels in the UI can update in response to the same change.