Remove babel polyfill and begin implemeting a lighter polyfill
This is primarily motivated by that the babel polyfill, by default,
shims just about everything, which makes it difficult to work through
which environment-defined functions (i.e. setTimeout) are natively
supported by Duktape. The babel polyfill also uses ES6 features which
are well optimized in browser engines but not so in Duktape. So this
begins a lightweight polyfill specifically for our use of Duktape in the
Blueprint application.