Initial version of update border drawing.
Minor adjustments to building the internal border-info property, and
then convert View to use that property and the data therein to draw
the border.

We build a separate path for each border edge and draw these
individually so a separate colour can be applied to each. Unfortunately
anti-aliasing at the end of a path drawn by JUCE means that it looks
like there's a small gap between lines. I don't have a good solution for
that at present.